agentic-api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pymongo
celery
redis
python-dotenv
passlib[bcrypt]
python-jose
pydantic
python-multipart
